requirements.txt
asttokens==3.0.0
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fsspec==2024.10.0
ghp-import==2.1.0
griffe==1.5.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.2.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.49
mkdocs-material-extensions==1.3.1
mkdocstrings==0.27.0
mkdocstrings-python==1.12.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.0
packaging==24.2
paginate==0.5.7
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
pymdown-extensions==10.12
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
